Refactor project structure and update dependencies
- Updated TypeScript definitions and devDependencies in package.json.
- Removed unused session.d.ts and AuthContext.tsx files.
- Refactored imports in layout.tsx, page.tsx, and other components to reflect new directory structure.
- Deleted legacy components related to watch list functionality.
- Updated MongoDB connection and authentication logic to use new paths.
- Cleaned up Tailwind configuration and adjusted TypeScript configuration for better compatibility.
