+++ app/lib/auth.js
