+++ tailwind.config.ts
@@ -1,4 +1,3 @@
-import type { Config } from 'tailwindcss'
 const withMT = require("@material-tailwind/react/utils/withMT");
 
 module.exports = withMT({
