+++ tsconfig.json
@@ -27,7 +27,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "components/WatchListButton/action.js"
+    "app/_components/WatchListButton/action.js"
   ],
   "exclude": ["node_modules"]
 }
